pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mycellm"
version = "0.2.3-dev"
description = "Distributed LLM inference across heterogeneous hardware"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Michael Gifford-Santos", email = "michael@mycellm.ai" },
]
keywords = ["llm", "distributed", "inference", "p2p"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "cryptography>=43.0",
    "aioquic>=1.0",
    "kademlia>=2.2",
    "cbor2>=5.6",
    "fastapi>=0.110",
    "uvicorn>=0.30",
    "sse-starlette>=2.0",
    "aiosqlite>=0.20",
    "sqlalchemy[asyncio]>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "prometheus-client>=0.20",
]

[project.optional-dependencies]
inference = [
    "llama-cpp-python>=0.3.19",
]
postgres = [
    "asyncpg>=0.29",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx>=0.27",
    "ruff",
]

[project.scripts]
mycellm = "mycellm.cli.main:app"

[project.urls]
Homepage = "https://mycellm.ai"
Documentation = "https://docs.mycellm.dev"
Repository = "https://repo.mycellm.dev/mycellm/mycellm"
Issues = "https://repo.mycellm.dev/mycellm/mycellm/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/mycellm"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "tests/",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
